+++ R/label.R
@@ -1,0 +1,5 @@
+label_cols <- function(mat, labels) {
+  dimnames(mat) <- list(NULL, labels)
+
+  mat
+}